\name{Demopack1-package}
\alias{Demopack1-package}
\alias{Demopack1}
\docType{package}
\title{
\packageTitle{Demopack1}
}
\description{
\packageDescription{Demopack1}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Demopack1}
\packageIndices{Demopack1}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Demopack1}

Maintainer: \packageMaintainer{Demopack1}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
